Updated URL name for showing items.
instead of calling the url `modelname.item`, call it `modelname.show` so
that it's consistent with the action that is called.

One funky thing about calling it show is that delete operations can be
done with the show URL, for example:

{% url "modelname.show" %}?_method=delete

I think the consistency is still a bit better.