Adds query map to ResourceView.
When the query params are being put together, reference the query map in
order to find the proper way to query for a particular field.  This
allows foreign keys to be properly queried and allows the user to
specify how a particular argument should be handled with respect to what
field within a model should be queried.